BACK-2704: Tests structure follows project's structure
