+++ test/cmd/config.test.js
@@ -14,11 +14,11 @@
  */
 
 const sinon = require('sinon');
-const command = require('./fixtures/command.js');
-const pkg = require('../package.json');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
-const config = require('../cmd/config.js');
+const command = require('../fixtures/command.js');
+const pkg = require('../../package.json');
+const project = require('../../lib/project.js');
+const user = require('../../lib/user.js');
+const config = require('../../cmd/config.js');
 
 describe(`./${pkg.name} config`, () => {
   const sandbox = sinon.sandbox.create();
