+++ test/cmd/job.test.js
@@ -14,14 +14,14 @@
  */
 
 const sinon = require('sinon');
-const command = require('./fixtures/command.js');
-const service = require('../lib/service.js');
-const logger = require('../lib/logger.js');
-const pkg = require('../package.json');
-const project = require('../lib/project.js');
-const job = require('../cmd/job.js');
-const user = require('../lib/user.js');
-const helper = require('./lib/helper');
+const command = require('../fixtures/command.js');
+const service = require('../../lib/service.js');
+const logger = require('../../lib/logger.js');
+const pkg = require('../../package.json');
+const project = require('../../lib/project.js');
+const job = require('../../cmd/job.js');
+const user = require('../../lib/user.js');
+const helper = require('../helper');
 
 describe(`./${pkg.name} job`, () => {
   const sandbox = sinon.sandbox.create();
