+++ test/cmd/list.test.js
@@ -14,38 +14,47 @@
  */
 
 const sinon = require('sinon');
-const command = require('./fixtures/command.js');
-const logout = require('../cmd/logout.js');
-const pkg = require('../package.json');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
+const command = require('../fixtures/command.js');
+const list = require('../../cmd/list.js');
+const logger = require('../../lib/logger.js');
+const pkg = require('../../package.json');
+const project = require('../../lib/project.js');
+const user = require('../../lib/user.js');
 
-describe(`./${pkg.name} logout`, () => {
+describe(`./${pkg.name} list`, () => {
   const sandbox = sinon.sandbox.create();
 
   before('setupStubs', () => {
-    sandbox.stub(user, 'logout').callsArg(0);
-    sandbox.stub(project, 'logout').callsArg(0);
+    sandbox.stub(user, 'setup').callsArg(1);
+    sandbox.stub(project, 'restore').callsArg(0);
+    sandbox.stub(project, 'list').callsArg(0);
   });
 
   afterEach('resetStubs', () => {
     sandbox.reset();
   });
 
-  after('user', () => {
+  after('cleanupStubs', () => {
     sandbox.restore();
   });
 
-  it('should logout the user.', (cb) => {
-    logout.call(command, command, (err) => {
-      expect(user.logout).to.be.calledOnce;
+  it('should setup the user.', (cb) => {
+    list.call(command, command, (err) => {
+      expect(user.setup).to.be.calledOnce;
       cb(err);
     });
   });
 
-  it('should logout the project.', (cb) => {
-    logout.call(command, command, (err) => {
-      expect(project.logout).to.be.calledOnce;
+  it('should restore the project.', (cb) => {
+    list.call(command, command, (err) => {
+      expect(project.restore).to.be.calledOnce;
+      cb(err);
+    });
+  });
+
+  it('should list the Kinvey datalinks.', (cb) => {
+    list.call(command, command, (err) => {
+      expect(project.list).to.be.calledOnce;
       cb(err);
     });
   });
