+++ test/cmd/recycle.test.js
@@ -14,20 +14,20 @@
  */
 
 const sinon = require('sinon');
-const command = require('./fixtures/command.js');
-const list = require('../cmd/list.js');
-const logger = require('../lib/logger.js');
-const pkg = require('../package.json');
-const project = require('../lib/project.js');
-const user = require('../lib/user.js');
+const command = require('../fixtures/command.js');
+const service = require('../../lib/service.js');
+const pkg = require('../../package.json');
+const project = require('../../lib/project.js');
+const recycle = require('../../cmd/recycle.js');
+const user = require('../../lib/user.js');
 
-describe(`./${pkg.name} list`, () => {
+describe(`./${pkg.name} recycle`, () => {
   const sandbox = sinon.sandbox.create();
 
   before('setupStubs', () => {
     sandbox.stub(user, 'setup').callsArg(1);
     sandbox.stub(project, 'restore').callsArg(0);
-    sandbox.stub(project, 'list').callsArg(0);
+    sandbox.stub(service, 'recycle').callsArg(0);
   });
 
   afterEach('resetStubs', () => {
@@ -39,22 +39,22 @@ describe(`./${pkg.name} list`, () => {
   });
 
   it('should setup the user.', (cb) => {
-    list.call(command, command, (err) => {
+    recycle.call(command, command, (err) => {
       expect(user.setup).to.be.calledOnce;
       cb(err);
     });
   });
 
   it('should restore the project.', (cb) => {
-    list.call(command, command, (err) => {
+    recycle.call(command, command, (err) => {
       expect(project.restore).to.be.calledOnce;
       cb(err);
     });
   });
 
-  it('should list the Kinvey datalinks.', (cb) => {
-    list.call(command, command, (err) => {
-      expect(project.list).to.be.calledOnce;
+  it('should reset the service.', (cb) => {
+    recycle.call(command, command, (err) => {
+      expect(service.recycle).to.be.calledOnce;
       cb(err);
     });
   });
